feat: allow to set a default project user level
