+++ apps/backend/src/database/util/user-level.ts
@@ -1,0 +1,6 @@
+export type UserLevel = "admin" | "reviewer" | "viewer";
+
+export const UserLevelJsonSchema = {
+  type: "string",
+  enum: ["admin", "reviewer", "viewer"],
+};
