+++ apps/frontend/src/gql/gql.ts
@@ -44,15 +44,17 @@ const documents = {
     "\n  fragment ProjectChangeName_Project on Project {\n    id\n    name\n    account {\n      id\n      slug\n    }\n  }\n": types.ProjectChangeName_ProjectFragmentDoc,
     "\n  mutation ProjectChangeName_updateProject($id: ID!, $name: String!) {\n    updateProject(input: { id: $id, name: $name }) {\n      id\n      name\n    }\n  }\n": types.ProjectChangeName_UpdateProjectDocument,
     "\n  query ConnectRepository($accountSlug: String!) {\n    account(slug: $accountSlug) {\n      __typename\n      id\n      gitlabAccessToken\n      glNamespaces {\n        edges {\n          id\n          kind\n          isProjectToken\n          ...GitlabNamespacesSelect_GlApiNamespace\n        }\n      }\n      permissions\n      ... on Team {\n        githubLightInstallation {\n          id\n          ghInstallation {\n            id\n            ...GithubInstallationsSelect_GhApiInstallation\n          }\n        }\n      }\n    }\n    me {\n      id\n      githubAccount {\n        id\n      }\n      ghInstallations {\n        edges {\n          id\n          ...GithubInstallationsSelect_GhApiInstallation\n        }\n        pageInfo {\n          totalCount\n        }\n      }\n    }\n  }\n": types.ConnectRepositoryDocument,
-    "\n  fragment ProjectContributedOnFragment on User {\n    projectsContributedOn(first: 1, projectId: $projectId) {\n      edges {\n        __typename\n        id\n        level\n      }\n    }\n  }\n": types.ProjectContributedOnFragmentFragmentDoc,
-    "\n  fragment ProjectContributors_Project on Project {\n    id\n    name\n    account {\n      id\n    }\n    permissions\n  }\n": types.ProjectContributors_ProjectFragmentDoc,
+    "\n  fragment ProjectContributors_Project on Project {\n    id\n    name\n    account {\n      id\n    }\n    permissions\n    defaultUserLevel\n    ...ProjectDefaultUserLevel_Project\n  }\n": types.ProjectContributors_ProjectFragmentDoc,
+    "\n  query ProjectContributors_TeamContributors(\n    $teamAccountId: ID!\n    $projectId: ID!\n    $search: String\n    $after: Int\n    $first: Int\n  ) {\n    team: teamById(id: $teamAccountId) {\n      id\n      members(\n        after: $after\n        first: $first\n        levels: [contributor]\n        search: $search\n      ) {\n        edges {\n          id\n          user {\n            id\n            ...ContributorListRow_user\n          }\n        }\n        pageInfo {\n          hasNextPage\n        }\n      }\n    }\n  }\n": types.ProjectContributors_TeamContributorsDocument,
+    "\n  fragment ContributorListRow_user on User {\n    id\n    projectsContributedOn(first: 1, projectId: $projectId) {\n      edges {\n        id\n        level\n      }\n    }\n    ...UserListRow_user\n  }\n": types.ContributorListRow_UserFragmentDoc,
+    "\n  mutation ProjectAddOrUpdateContributorMutation(\n    $projectId: ID!\n    $userAccountId: ID!\n    $level: ProjectUserLevel!\n  ) {\n    addOrUpdateProjectContributor(\n      input: {\n        projectId: $projectId\n        userAccountId: $userAccountId\n        level: $level\n      }\n    ) {\n      id\n      level\n    }\n  }\n": types.ProjectAddOrUpdateContributorMutationDocument,
     "\n  query ProjectContributorsQuery($projectId: ID!, $after: Int, $first: Int) {\n    project: projectById(id: $projectId) {\n      id\n      contributors(after: $after, first: $first) {\n        edges {\n          id\n          level\n          user {\n            id\n            ...UserListRow_user\n          }\n        }\n        pageInfo {\n          hasNextPage\n        }\n      }\n    }\n  }\n": types.ProjectContributorsQueryDocument,
+    "\n  fragment ProjectDefaultUserLevel_Project on Project {\n    id\n    defaultUserLevel\n  }\n": types.ProjectDefaultUserLevel_ProjectFragmentDoc,
+    "\n  mutation ProjectDefaultUserLevel_updateProject(\n    $id: ID!\n    $defaultUserLevel: ProjectUserLevel\n  ) {\n    updateProject(input: { id: $id, defaultUserLevel: $defaultUserLevel }) {\n      id\n      defaultUserLevel\n    }\n  }\n": types.ProjectDefaultUserLevel_UpdateProjectDocument,
     "\n  mutation RemoveContributorFromProjectMutation(\n    $projectId: ID!\n    $userAccountId: ID!\n  ) {\n    removeContributorFromProject(\n      input: { projectId: $projectId, userAccountId: $userAccountId }\n    ) {\n      projectContributorId\n    }\n  }\n": types.RemoveContributorFromProjectMutationDocument,
     "\n  fragment RemoveFromProjectDialog_User on User {\n    id\n    ...UserListRow_user\n  }\n": types.RemoveFromProjectDialog_UserFragmentDoc,
-    "\n  query ProjectContributors_TeamContributors(\n    $teamAccountId: ID!\n    $projectId: ID!\n    $search: String\n    $after: Int\n    $first: Int\n  ) {\n    team: teamById(id: $teamAccountId) {\n      id\n      members(\n        after: $after\n        first: $first\n        levels: [contributor]\n        search: $search\n      ) {\n        edges {\n          id\n          user {\n            id\n            ...ContributorListRow_user\n          }\n        }\n        pageInfo {\n          hasNextPage\n        }\n      }\n    }\n  }\n": types.ProjectContributors_TeamContributorsDocument,
-    "\n  mutation ProjectAddOrUpdateContributorMutation(\n    $projectId: ID!\n    $userAccountId: ID!\n    $level: ProjectUserLevel!\n  ) {\n    addOrUpdateProjectContributor(\n      input: {\n        projectId: $projectId\n        userAccountId: $userAccountId\n        level: $level\n      }\n    ) {\n      id\n      level\n    }\n  }\n": types.ProjectAddOrUpdateContributorMutationDocument,
-    "\n  fragment ContributorListRow_user on User {\n    id\n    projectsContributedOn(first: 1, projectId: $projectId) {\n      edges {\n        id\n        level\n      }\n    }\n    ...UserListRow_user\n  }\n": types.ContributorListRow_UserFragmentDoc,
     "\n  query ProjectContributors_TeamMembers(\n    $teamAccountId: ID!\n    $search: String\n    $after: Int\n    $first: Int\n  ) {\n    team: teamById(id: $teamAccountId) {\n      id\n      members(\n        after: $after\n        first: $first\n        levels: [owner, member]\n        search: $search\n      ) {\n        edges {\n          id\n          level\n          user {\n            id\n            ...UserListRow_user\n          }\n        }\n        pageInfo {\n          hasNextPage\n        }\n      }\n    }\n  }\n": types.ProjectContributors_TeamMembersDocument,
+    "\n  fragment ProjectContributedOnFragment on User {\n    projectsContributedOn(first: 1, projectId: $projectId) {\n      edges {\n        __typename\n        id\n        level\n      }\n    }\n  }\n": types.ProjectContributedOnFragmentFragmentDoc,
     "\n  mutation DeleteProjectMutation($projectId: ID!) {\n    deleteProject(id: $projectId)\n  }\n": types.DeleteProjectMutationDocument,
     "\n  fragment ProjectDelete_Project on Project {\n    id\n    name\n    account {\n      id\n      slug\n    }\n  }\n": types.ProjectDelete_ProjectFragmentDoc,
     "\n  fragment ProjectGitRepository_Project on Project {\n    id\n    account {\n      id\n      slug\n    }\n    repository {\n      __typename\n      id\n      fullName\n      url\n    }\n    prCommentEnabled\n  }\n": types.ProjectGitRepository_ProjectFragmentDoc,
@@ -276,15 +278,31 @@ export function graphql(source: "\n  query ConnectRepository($accountSlug: Strin
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
-export function graphql(source: "\n  fragment ProjectContributedOnFragment on User {\n    projectsContributedOn(first: 1, projectId: $projectId) {\n      edges {\n        __typename\n        id\n        level\n      }\n    }\n  }\n"): (typeof documents)["\n  fragment ProjectContributedOnFragment on User {\n    projectsContributedOn(first: 1, projectId: $projectId) {\n      edges {\n        __typename\n        id\n        level\n      }\n    }\n  }\n"];
+export function graphql(source: "\n  fragment ProjectContributors_Project on Project {\n    id\n    name\n    account {\n      id\n    }\n    permissions\n    defaultUserLevel\n    ...ProjectDefaultUserLevel_Project\n  }\n"): (typeof documents)["\n  fragment ProjectContributors_Project on Project {\n    id\n    name\n    account {\n      id\n    }\n    permissions\n    defaultUserLevel\n    ...ProjectDefaultUserLevel_Project\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
-export function graphql(source: "\n  fragment ProjectContributors_Project on Project {\n    id\n    name\n    account {\n      id\n    }\n    permissions\n  }\n"): (typeof documents)["\n  fragment ProjectContributors_Project on Project {\n    id\n    name\n    account {\n      id\n    }\n    permissions\n  }\n"];
+export function graphql(source: "\n  query ProjectContributors_TeamContributors(\n    $teamAccountId: ID!\n    $projectId: ID!\n    $search: String\n    $after: Int\n    $first: Int\n  ) {\n    team: teamById(id: $teamAccountId) {\n      id\n      members(\n        after: $after\n        first: $first\n        levels: [contributor]\n        search: $search\n      ) {\n        edges {\n          id\n          user {\n            id\n            ...ContributorListRow_user\n          }\n        }\n        pageInfo {\n          hasNextPage\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query ProjectContributors_TeamContributors(\n    $teamAccountId: ID!\n    $projectId: ID!\n    $search: String\n    $after: Int\n    $first: Int\n  ) {\n    team: teamById(id: $teamAccountId) {\n      id\n      members(\n        after: $after\n        first: $first\n        levels: [contributor]\n        search: $search\n      ) {\n        edges {\n          id\n          user {\n            id\n            ...ContributorListRow_user\n          }\n        }\n        pageInfo {\n          hasNextPage\n        }\n      }\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  fragment ContributorListRow_user on User {\n    id\n    projectsContributedOn(first: 1, projectId: $projectId) {\n      edges {\n        id\n        level\n      }\n    }\n    ...UserListRow_user\n  }\n"): (typeof documents)["\n  fragment ContributorListRow_user on User {\n    id\n    projectsContributedOn(first: 1, projectId: $projectId) {\n      edges {\n        id\n        level\n      }\n    }\n    ...UserListRow_user\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  mutation ProjectAddOrUpdateContributorMutation(\n    $projectId: ID!\n    $userAccountId: ID!\n    $level: ProjectUserLevel!\n  ) {\n    addOrUpdateProjectContributor(\n      input: {\n        projectId: $projectId\n        userAccountId: $userAccountId\n        level: $level\n      }\n    ) {\n      id\n      level\n    }\n  }\n"): (typeof documents)["\n  mutation ProjectAddOrUpdateContributorMutation(\n    $projectId: ID!\n    $userAccountId: ID!\n    $level: ProjectUserLevel!\n  ) {\n    addOrUpdateProjectContributor(\n      input: {\n        projectId: $projectId\n        userAccountId: $userAccountId\n        level: $level\n      }\n    ) {\n      id\n      level\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(source: "\n  query ProjectContributorsQuery($projectId: ID!, $after: Int, $first: Int) {\n    project: projectById(id: $projectId) {\n      id\n      contributors(after: $after, first: $first) {\n        edges {\n          id\n          level\n          user {\n            id\n            ...UserListRow_user\n          }\n        }\n        pageInfo {\n          hasNextPage\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query ProjectContributorsQuery($projectId: ID!, $after: Int, $first: Int) {\n    project: projectById(id: $projectId) {\n      id\n      contributors(after: $after, first: $first) {\n        edges {\n          id\n          level\n          user {\n            id\n            ...UserListRow_user\n          }\n        }\n        pageInfo {\n          hasNextPage\n        }\n      }\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  fragment ProjectDefaultUserLevel_Project on Project {\n    id\n    defaultUserLevel\n  }\n"): (typeof documents)["\n  fragment ProjectDefaultUserLevel_Project on Project {\n    id\n    defaultUserLevel\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  mutation ProjectDefaultUserLevel_updateProject(\n    $id: ID!\n    $defaultUserLevel: ProjectUserLevel\n  ) {\n    updateProject(input: { id: $id, defaultUserLevel: $defaultUserLevel }) {\n      id\n      defaultUserLevel\n    }\n  }\n"): (typeof documents)["\n  mutation ProjectDefaultUserLevel_updateProject(\n    $id: ID!\n    $defaultUserLevel: ProjectUserLevel\n  ) {\n    updateProject(input: { id: $id, defaultUserLevel: $defaultUserLevel }) {\n      id\n      defaultUserLevel\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
@@ -296,19 +314,11 @@ export function graphql(source: "\n  fragment RemoveFromProjectDialog_User on Us
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
-export function graphql(source: "\n  query ProjectContributors_TeamContributors(\n    $teamAccountId: ID!\n    $projectId: ID!\n    $search: String\n    $after: Int\n    $first: Int\n  ) {\n    team: teamById(id: $teamAccountId) {\n      id\n      members(\n        after: $after\n        first: $first\n        levels: [contributor]\n        search: $search\n      ) {\n        edges {\n          id\n          user {\n            id\n            ...ContributorListRow_user\n          }\n        }\n        pageInfo {\n          hasNextPage\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query ProjectContributors_TeamContributors(\n    $teamAccountId: ID!\n    $projectId: ID!\n    $search: String\n    $after: Int\n    $first: Int\n  ) {\n    team: teamById(id: $teamAccountId) {\n      id\n      members(\n        after: $after\n        first: $first\n        levels: [contributor]\n        search: $search\n      ) {\n        edges {\n          id\n          user {\n            id\n            ...ContributorListRow_user\n          }\n        }\n        pageInfo {\n          hasNextPage\n        }\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  mutation ProjectAddOrUpdateContributorMutation(\n    $projectId: ID!\n    $userAccountId: ID!\n    $level: ProjectUserLevel!\n  ) {\n    addOrUpdateProjectContributor(\n      input: {\n        projectId: $projectId\n        userAccountId: $userAccountId\n        level: $level\n      }\n    ) {\n      id\n      level\n    }\n  }\n"): (typeof documents)["\n  mutation ProjectAddOrUpdateContributorMutation(\n    $projectId: ID!\n    $userAccountId: ID!\n    $level: ProjectUserLevel!\n  ) {\n    addOrUpdateProjectContributor(\n      input: {\n        projectId: $projectId\n        userAccountId: $userAccountId\n        level: $level\n      }\n    ) {\n      id\n      level\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  fragment ContributorListRow_user on User {\n    id\n    projectsContributedOn(first: 1, projectId: $projectId) {\n      edges {\n        id\n        level\n      }\n    }\n    ...UserListRow_user\n  }\n"): (typeof documents)["\n  fragment ContributorListRow_user on User {\n    id\n    projectsContributedOn(first: 1, projectId: $projectId) {\n      edges {\n        id\n        level\n      }\n    }\n    ...UserListRow_user\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
 export function graphql(source: "\n  query ProjectContributors_TeamMembers(\n    $teamAccountId: ID!\n    $search: String\n    $after: Int\n    $first: Int\n  ) {\n    team: teamById(id: $teamAccountId) {\n      id\n      members(\n        after: $after\n        first: $first\n        levels: [owner, member]\n        search: $search\n      ) {\n        edges {\n          id\n          level\n          user {\n            id\n            ...UserListRow_user\n          }\n        }\n        pageInfo {\n          hasNextPage\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query ProjectContributors_TeamMembers(\n    $teamAccountId: ID!\n    $search: String\n    $after: Int\n    $first: Int\n  ) {\n    team: teamById(id: $teamAccountId) {\n      id\n      members(\n        after: $after\n        first: $first\n        levels: [owner, member]\n        search: $search\n      ) {\n        edges {\n          id\n          level\n          user {\n            id\n            ...UserListRow_user\n          }\n        }\n        pageInfo {\n          hasNextPage\n        }\n      }\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  fragment ProjectContributedOnFragment on User {\n    projectsContributedOn(first: 1, projectId: $projectId) {\n      edges {\n        __typename\n        id\n        level\n      }\n    }\n  }\n"): (typeof documents)["\n  fragment ProjectContributedOnFragment on User {\n    projectsContributedOn(first: 1, projectId: $projectId) {\n      edges {\n        __typename\n        id\n        level\n      }\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
